pyproject.toml
[tool.poetry]
name = "langfuse-v2"
version = "0.1.0"
description = ""
authors = ["xiang.yu <xiang.yu@jera.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "*"
pytz = "*"
langfuse = "*"
python-dotenv = "*"
azure-storage-blob = "*"
psycopg2 = "*"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
